finish refactoring, tests are missing
