+++ src/main/webapp/app/overview/course-overview.component.html
@@ -23,8 +23,6 @@
                     [courses]="courses"
                     [sidebarItems]="sidebarItems"
                     [courseActionItems]="courseActionItems"
-                    [hiddenItems]="hiddenItems"
-                    [anyItemHidden]="anyItemHidden"
                     [isNavbarCollapsed]="isNavbarCollapsed"
                     [isExamStarted]="isExamStarted"
                     [isProduction]="isProduction"
